+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 Version 0.8.1 (UNRELEASED)
 --------------------------
 
+- Adds support for creating reana-client standalone AppImage executables.
 - Adds support for Python 3.10.
 - Adds workflow name validation for ``create_workflow_from_json()`` Python API function.